fix start bash bug

+++ run_recover.sh
@@ -1,3 +1,8 @@
+  if [ -f script.pid ]; then
+    echo "recover already exist."
+    exit 0;
+  fi
+  
   echo $$ > script.pid
 
   ./target/release/recover_state --genesis >> log/recover_state.log 2>&1 &
@@ -10,6 +15,7 @@
   cd ..
 
   wait $recover_state_pid
+  rm script.pid
 
   nohup ./target/release/exodus_server >> log/server.log 2>&1 &
   nohup ./target/release/exodus_prover tasks -w 1 >> log/prover.log 2>&1 &
